FinetuneData/face_requirements.txt
opencv-python>=4.8.0
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.65.0
